abstrae metodo grafico a su propio archivo
